bug fix for loading of xsl stylesheet from string
The openbibl/saxon module had a bug handling an xsl transform
starting with an xml-string of an xsl stylesheet, where the
stylesheet contained xsl:import statements. The processor
produced this error message

   SEVERE: Error at /xsl:stylesheet/xsl:import
   Cannot initialize URI with empty parameters

resulting in this failure

   SEVERE: XPathException in renderXML: Failed to
   compile stylesheet. 1 error detected.

The fix is to manually set the @xml:base attribute
and value for the xsl:stylesheet element immediately
after it was parsed to a Document.