added typeahead config to Jasmine test config
The Bootstrap Typeahead extension to the global jQuery
object sometimes (but not always) casues an error to
be thrown due to an undefined $ variable, with the error
(in Chrome):

  Uncaught TypeError: Cannot read property 'fn' of
  undefined

This update to the config is meant to ensure that
the jQuery global is defined when the Typeahead
module is loaded.

These errors happen only when the tests are run
remotely (from *.github.io domain).