multiple
Numerous changes in the JS and XSL code around:

1) Improvement in project pathing determinations, from
xsl/openbibl.boot.xsl to initial configuration in openbibl/
state.js. Led to numerous changes in Jasmine tests, but a
more consistent file-path-retrieval scheme.

2) Removal of HTML partial-view templates, in favor of
using Saxon-CE/XSLT for targeted updates to the DOM.
Updated Saxon-CE calling convention from within the obp
code at the same time, to simplify. Updated the "download
to html" code to not remove the Saxon-CE js library call,
since it is now used in partial-view updates.

3) Addition of link in tooltip text, pointing to related
entries with a mention of the same entity. Included
additions to the LESS/CSS config for new tooltip elements.

Commit also included updates to documentation, and
updates to the Jasmine tests as mentioned above. Jasmine
tests pass, but one tooltip-related test fails intermittently
so it will need to be revisited.